refactor: abstract component creation to a styled function
Also update snaps according to our latest changes
