+++ src/components/AbstractButton/AbstractButton.tsx
@@ -1,15 +1,23 @@
-import Box, { BoxProps } from '../Box';
-import { defaultButtonStyles } from './utils';
+import React from 'react';
+import { pounce } from '../../system';
+import { __DEV__ } from '../../utils/helpers';
 
-export type AbstractButtonProps = BoxProps<'button'>;
+export type AbstractButtonProps = React.ComponentProps<typeof AbstractButton>;
 
-export const AbstractButton = Box;
+export const AbstractButton = pounce('button');
 
 // The component should render a button by default, with some default styles
 AbstractButton.defaultProps = {
-  as: 'button',
   type: 'button',
-  ...defaultButtonStyles,
+  cursor: 'pointer',
+  color: 'gray-50' as const,
+  textDecoration: 'none',
+  backgroundColor: 'transparent' as const,
+  transition: 'all 0.1s linear',
 };
 
 export default AbstractButton;
+
+if (__DEV__) {
+  AbstractButton.displayName = 'AbstractButton';
+}
