+++ src/system/index.ts
@@ -2,3 +2,4 @@ export * from './system';
 export * from './utility';
 export * from './pseudo';
 export * from './shouldForwardProp';
+export * from './styled';
